refactor(backend): split OpenAI loop internals
